Move some more code from rifServices to the new statsService module.
